move cert-related code to dedicated source file
